implemented dropdown selection & mapped siteID to name
